requirements.txt
argparse
pan-os-python
pyyaml
xmltodict
ipaddress